Fixed LocalVariable tables in Soot for parameters and 'this' to always start on the first unit of the body of a method. This fixes an issue with the LocalSplitter splitting parameters which are written without being read first. (#597)
Also removed the special compilation of Object.<init>() as it didn't create
allocas as expected by compiler plugins. Object.<init>() will now be compiled
as any other method and the call to register_finalizable() will be inserted
afterwards.

Also fixed so that the nop inserted by MethodCompiler to workaround a bug in
Soot's ArrayBoundsCheckerAnalysis (see #1) is removed right after the Soot
passes have been run. Otherwise the start unit for parameter LocalVariables
will not be the first unit of the method any longer.